requirements.txt
fastapi
uvicorn
python-dotenv
google-genai
pinecone-client
cohere
langchain
langchain-community
pypdf